--- lumen/ai/controls/ingest/api.py
@@ -1,5 +1,0 @@
-# Backward-compatibility re-exports.  Import from rest_api / open_api directly.
-from .open_api import OpenAPISourceControls
-from .rest_api import RESTAPISourceControls
-
-__all__ = ("OpenAPISourceControls", "RESTAPISourceControls")
